Migrate search to use useSearch hooks
