+++ src/pages/search/index.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState, useRef, ChangeEvent, useMemo } from "react";
 import { useRouter } from "next/router";
 import { useTranslation } from "react-i18next";
-import { TDocument, TFamily } from "@types";
-import { getSearch } from "@hooks/useSearch";
+import { TDocument } from "@types";
+import useSearch from "@hooks/useSearch";
 import useDocument from "@hooks/useDocument";
 import useUpdateDocument from "@hooks/useUpdateDocument";
 import useUpdateCountries from "@hooks/useUpdateCountries";
@@ -31,11 +31,11 @@ import buildSearchQuery from "@utils/buildSearchQuery";
 import { PER_PAGE } from "@constants/paging";
 import { DOCUMENT_CATEGORIES } from "@constants/documentCategories";
 import { QUERY_PARAMS } from "@constants/queryParams";
-import { getCachedSearch, updateCacheSearch, TCacheResult } from "@utils/searchCache";
 
 const Search = () => {
   const router = useRouter();
   const qQueryString = router.query[QUERY_PARAMS.query_string];
+  const isBrowsing = !qQueryString || qQueryString?.toString().trim() === "";
   const slideoutRef = useRef(null);
   const { t, ready } = useTranslation(["searchStart", "searchResults"]);
   const [showFilters, setShowFilters] = useState(false);
@@ -43,16 +43,9 @@ const Search = () => {
   const [showPDF, setShowPDF] = useState(false);
   const [passageIndex, setPassageIndex] = useState(null);
   const [pageCount, setPageCount] = useState(1);
-  const [isLoading, setIsLoading] = useState(false);
-  const [families, setFamilies] = useState<TFamily[]>([]);
-  const [hits, setHits] = useState<number>(null);
 
   const updateDocument = useUpdateDocument();
   const updateCountries = useUpdateCountries();
-
-  const searchQuery = useMemo(() => {
-    return buildSearchQuery({ ...router.query });
-  }, [router.query]);
 
   // close slideout panel when clicking outside of it
   useOutsideAlerter(slideoutRef, (e) => {
@@ -62,19 +55,16 @@ const Search = () => {
     setShowSlideout(false);
   });
 
+  const searchQuery = useMemo(() => {
+    return buildSearchQuery({ ...router.query });
+  }, [router.query]);
+
+  const { status, families, hits } = useSearch(searchQuery);
+
   const configQuery = useConfig("config");
   const { data: { document_types = [], sectors = [], regions = [], countries = [] } = {} } = configQuery;
 
   const { data: filteredCountries } = useFilteredCountries(countries);
-
-  const isBrowsing = !qQueryString || qQueryString?.toString().trim() === "";
-
-  // FIXME: remove this
-  // const resultsQuery = useSearch("searches", searchQuery);
-  // const families = resultsQuery?.data?.data?.families ?? [];
-  // const hits = resultsQuery?.data?.data?.hits ?? 0;
-  // const families = [];
-  // const hits = 0;
 
   const { data: document }: { data: TDocument } = ({} = useDocument());
 
@@ -243,44 +233,6 @@ const Search = () => {
     }
   }, [hits]);
 
-  useEffect(() => {
-    setIsLoading(true);
-    // Check if we have a cached result before calling the API
-    const cacheId = {
-      query_string: searchQuery.query_string,
-      exact_match: searchQuery.exact_match,
-      keyword_filters: searchQuery.keyword_filters,
-      year_range: searchQuery.year_range,
-      sort_field: searchQuery.sort_field,
-      sort_order: searchQuery.sort_order,
-      offset: searchQuery.offset,
-    };
-    const cachedResult = getCachedSearch(cacheId);
-    if (cachedResult) {
-      setFamilies(cachedResult?.families || []);
-      setHits(cachedResult.hits);
-      setIsLoading(false);
-      return;
-    }
-
-    const resultsQuery = getSearch(searchQuery);
-    resultsQuery.then((res) => {
-      if (res.status === 200) {
-        setFamilies(res.data.families);
-        setHits(res.data.hits);
-
-        const searchToCache: TCacheResult = {
-          ...cacheId,
-          families: res.data.families,
-          hits: res.data.hits,
-          timestamp: new Date().getTime(),
-        };
-        updateCacheSearch(searchToCache);
-      }
-      setIsLoading(false);
-    });
-  }, [searchQuery]);
-
   const renderNoOfResults = () => {
     let resultsMsg = `Showing`;
     if (hits < 100) {
@@ -381,7 +333,7 @@ const Search = () => {
                 <div className="md:w-3/4">
                   <div className="md:pl-8">
                     <div className="lg:flex justify-between">
-                      <div className="text-sm my-4 md:my-0 md:flex">{!isLoading && renderNoOfResults()}</div>
+                      <div className="text-sm my-4 md:my-0 md:flex">{status === "fetched" && renderNoOfResults()}</div>
                       <ExternalLink
                         url="https://docs.google.com/forms/d/e/1FAIpQLSdFkgTNfzms7PCpfIY3d2xGDP5bYXx8T2-2rAk_BOmHMXvCoA/viewform"
                         className="text-sm text-blue-600 mt-4 md:mt-0 hover:underline"
@@ -400,7 +352,7 @@ const Search = () => {
                   </div>
 
                   <div className="search-results md:pl-8 md:mt-12 relative">
-                    {isLoading ? (
+                    {status === "loading" ? (
                       <div className="w-full flex justify-center h-96">
                         <Loader />
                       </div>
